refactor: перенёс статистику в "статистику" + перенёс локали
